// clients/src/utilities/styles.tsx
import styled from "styled-components";

export const DefaultContainer = styled.div`
    flex: 0 0 20%;

    h4 {
        margin: 6px 0;
        color: #ffffff;
        font-size: 14px;
        font-weight: normal;
    }
`;

export const DefaultBtn = styled.div`
    button {
        padding: 14px;
        background: transparent;
        border-radius: 5px;
        font-weight: 600;
        display: flex;
        margin-left: auto;
        font-size: 16px;

        svg {
            padding-right: 16px;
        }
    }
`